fix: fit transform should not fit at encode time plus refactor
